+++ src/api/noteapi.h
@@ -13,6 +13,7 @@ class NoteApi : public QObject, public Note {
     Q_PROPERTY(QString name READ getName() WRITE setName)
     Q_PROPERTY(QString fileName READ getFileName())
     Q_PROPERTY(QString fullNoteFilePath READ fullNoteFilePath())
+    Q_PROPERTY(QString fullNoteFileDirPath READ fullNoteFileDirPath())
     Q_PROPERTY(int noteSubFolderId READ getNoteSubFolderId())
     Q_PROPERTY(QString noteText READ getNoteText() WRITE setNoteText())
     Q_PROPERTY(QString decryptedNoteText READ getDecryptedNoteText()
@@ -38,4 +39,5 @@ public:
     Q_INVOKABLE QQmlListProperty<NoteApi> fetchAll(int limit = -1,
                                                    int offset = -1);
     Q_INVOKABLE QString toMarkdownHtml(bool forExport = true);
+    Q_INVOKABLE const QString getFileURLFromFileName(QString localFileName);
 };